+++ src/main.c
@@ -44,6 +44,9 @@ void help_message(Command command)
         case COMMAND_GENERATE:
             generate_help();
             break;
+        case COMMAND_IMPORT:
+            import_help();
+            break;
         case COMMAND_UNSET:
         case COMMAND_UNKNOWN:
         case COMMAND_INIT: